Restore symlink following feature not ported from original deno version
If you create a relative symlink inside the home directory, to something
else in your dotfiles directory, before link_o_matic would create a
symlink to that symlink which the OS would resolve.

Now the symlink is preresolved and the link that is created is to the
target of the intermediate symlink, as described in the Readme file.

This feature was not originally ported from the Deno version when the
Rust version was developed.